+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/DataMovementExtensions.cs
@@ -23,13 +23,10 @@ namespace Azure.Storage.DataMovement
 
         public static StreamToUriJobPart ToStreamToUriJobPartAsync(
             this TransferJobInternal baseJob,
-            Stream planFileStream,
+            JobPartPlanHeader header,
             StorageResourceItem sourceResource,
             StorageResourceItem destinationResource)
         {
-            // Convert stream to job plan header
-            JobPartPlanHeader header = JobPartPlanHeader.Deserialize(planFileStream);
-
             // Override header values if options were specified by user.
             long initialTransferSize = baseJob._initialTransferSize ?? header.InitialTransferSize;
             long transferChunkSize = baseJob._maximumTransferChunkSize ?? header.ChunkSize;
@@ -55,13 +52,10 @@ namespace Azure.Storage.DataMovement
 
         public static ServiceToServiceJobPart ToServiceToServiceJobPartAsync(
             this TransferJobInternal baseJob,
-            Stream planFileStream,
+            JobPartPlanHeader header,
             StorageResourceItem sourceResource,
             StorageResourceItem destinationResource)
         {
-            // Convert stream to job plan header
-            JobPartPlanHeader header = JobPartPlanHeader.Deserialize(planFileStream);
-
             // Override header values if options were specified by user.
             long initialTransferSize = baseJob._initialTransferSize ?? header.InitialTransferSize;
             long transferChunkSize = baseJob._maximumTransferChunkSize ?? header.ChunkSize;
@@ -87,13 +81,10 @@ namespace Azure.Storage.DataMovement
 
         public static UriToStreamJobPart ToUriToStreamJobPartAsync(
             this TransferJobInternal baseJob,
-            Stream planFileStream,
+            JobPartPlanHeader header,
             StorageResourceItem sourceResource,
             StorageResourceItem destinationResource)
         {
-            // Convert stream to job plan header
-            JobPartPlanHeader header = JobPartPlanHeader.Deserialize(planFileStream);
-
             // Override header values if options were specified by user.
             long initialTransferSize = baseJob._initialTransferSize ?? header.InitialTransferSize;
             long transferChunkSize = baseJob._maximumTransferChunkSize ?? header.ChunkSize;
@@ -119,13 +110,10 @@ namespace Azure.Storage.DataMovement
 
         public static StreamToUriJobPart ToStreamToUriJobPartAsync(
             this TransferJobInternal baseJob,
-            Stream planFileStream,
+            JobPartPlanHeader header,
             StorageResourceContainer sourceResource,
             StorageResourceContainer destinationResource)
         {
-            // Convert stream to job plan header
-            JobPartPlanHeader header = JobPartPlanHeader.Deserialize(planFileStream);
-
             string childSourcePath = header.SourcePath;
             string childSourceName = childSourcePath.Substring(sourceResource.Uri.AbsoluteUri.Length + 1);
             string childDestinationPath = header.DestinationPath;
@@ -155,13 +143,10 @@ namespace Azure.Storage.DataMovement
 
         public static ServiceToServiceJobPart ToServiceToServiceJobPartAsync(
             this TransferJobInternal baseJob,
-            Stream planFileStream,
+            JobPartPlanHeader header,
             StorageResourceContainer sourceResource,
             StorageResourceContainer destinationResource)
         {
-            // Convert stream to job plan header
-            JobPartPlanHeader header = JobPartPlanHeader.Deserialize(planFileStream);
-
             string childSourcePath = header.SourcePath;
             string childDestinationPath = header.DestinationPath;
             // Override header values if options were specified by user.
@@ -189,13 +174,10 @@ namespace Azure.Storage.DataMovement
 
         public static UriToStreamJobPart ToUriToStreamJobPartAsync(
             this TransferJobInternal baseJob,
-            Stream planFileStream,
+            JobPartPlanHeader header,
             StorageResourceContainer sourceResource,
             StorageResourceContainer destinationResource)
         {
-            // Convert stream to job plan header
-            JobPartPlanHeader header = JobPartPlanHeader.Deserialize(planFileStream);
-
             // Apply credentials to the saved transfer job path
             string childSourcePath = header.SourcePath;
             string childSourceName = childSourcePath.Substring(sourceResource.Uri.AbsoluteUri.Length + 1);